viz-selector now closes every time
