+++ frontend/src/metabase/dashboard/components/DashCard/DashCard.tsx
@@ -1,13 +1,16 @@
+import { useDisclosure } from "@mantine/hooks";
 import cx from "classnames";
 import type { LocationDescriptor } from "history";
 import { getIn } from "icepick";
 import { memo, useCallback, useMemo, useRef, useState } from "react";
 import { useMount, useUpdateEffect } from "react-use";
+import { t } from "ttag";
 
 import ErrorBoundary from "metabase/ErrorBoundary";
 import { isActionCard } from "metabase/actions/utils";
 import CS from "metabase/css/core/index.css";
 import DashboardS from "metabase/css/dashboard.module.css";
+import { replaceCardWithVisualization } from "metabase/dashboard/actions";
 import { DASHBOARD_SLOW_TIMEOUT } from "metabase/dashboard/constants";
 import { getDashcardData, getDashcardHref } from "metabase/dashboard/selectors";
 import {
@@ -17,13 +20,19 @@ import {
 } from "metabase/dashboard/utils";
 import { isEmbeddingSdk } from "metabase/env";
 import { color } from "metabase/lib/colors";
-import { useSelector, useStore } from "metabase/lib/redux";
+import { useDispatch, useSelector, useStore } from "metabase/lib/redux";
 import { PLUGIN_COLLECTIONS } from "metabase/plugins";
 import EmbedFrameS from "metabase/public/components/EmbedFrame/EmbedFrame.module.css";
 import { Box } from "metabase/ui";
 import { getVisualizationRaw } from "metabase/visualizations";
 import { extendCardWithDashcardSettings } from "metabase/visualizations/lib/settings/typed-utils";
 import type { ClickActionModeGetter } from "metabase/visualizations/types";
+import { VisualizerModal } from "metabase/visualizer/components/VisualizerModal";
+import {
+  dashboardCardSupportsVisualizer,
+  getInitialStateForCardDataSource,
+  isVisualizerDashboardCard,
+} from "metabase/visualizer/utils";
 import type {
   Card,
   CardId,
@@ -34,6 +43,7 @@ import type {
   VisualizationSettings,
 } from "metabase-types/api";
 import type { StoreDashcard } from "metabase-types/store";
+import type { VisualizerHistoryItem } from "metabase-types/store/visualizer";
 
 import S from "./DashCard.module.css";
 import { DashCardActionsPanel } from "./DashCardActionsPanel/DashCardActionsPanel";
@@ -132,6 +142,8 @@ function DashCardInner({
   editDashboard,
   className,
 }: DashCardProps) {
+  const dispatch = useDispatch();
+
   const dashcardData = useSelector(state =>
     getDashcardData(state, dashcard.id),
   );
@@ -307,6 +319,54 @@ function DashCardInner({
       [dashcard, navigateToNewCardFromDashboard],
     );
 
+  const [
+    isVisualizerModalOpen,
+    { open: openVisualizerModal, close: closeVisualizerModal },
+  ] = useDisclosure(false);
+
+  const onEditVisualization = useMemo(() => {
+    if (dashcard && dashboardCardSupportsVisualizer(dashcard)) {
+      return openVisualizerModal;
+    }
+  }, [dashcard, openVisualizerModal]);
+
+  const onVisualizerModalSave = useCallback(
+    (visualization: VisualizerHistoryItem) => {
+      dispatch(
+        replaceCardWithVisualization({
+          dashcardId: dashcard.id,
+          visualization,
+        }),
+      );
+      closeVisualizerModal();
+    },
+    [dashcard.id, dispatch, closeVisualizerModal],
+  );
+
+  const onVisualizerModalClose = useCallback(() => {
+    closeVisualizerModal();
+  }, [closeVisualizerModal]);
+
+  const visualizerModalInitialState = useMemo(() => {
+    if (!isVisualizerModalOpen) {
+      return;
+    }
+
+    if (isVisualizerDashboardCard(dashcard)) {
+      return {
+        state: dashcard.visualization_settings
+          ?.visualization as Partial<VisualizerHistoryItem>,
+      };
+    } else {
+      return {
+        state: getInitialStateForCardDataSource(
+          series[0].card,
+          series[0].data?.cols ?? [],
+        ),
+      };
+    }
+  }, [dashcard, series, isVisualizerModalOpen]);
+
   return (
     <ErrorBoundary>
       <Box
@@ -366,6 +426,7 @@ function DashCardInner({
             showClickBehaviorSidebar={handleShowClickBehaviorSidebar}
             onPreviewToggle={handlePreviewToggle}
             isTrashedOnRemove={isTrashedOnRemove}
+            onEditVisualization={onEditVisualization}
           />
         )}
         <DashCardVisualization
@@ -403,8 +464,17 @@ function DashCardInner({
           onTogglePreviewing={handlePreviewToggle}
           downloadsEnabled={downloadsEnabled}
           editDashboard={editDashboard}
+          onEditVisualization={onEditVisualization}
         />
       </Box>
+      {isVisualizerModalOpen && (
+        <VisualizerModal
+          onSave={onVisualizerModalSave}
+          onClose={onVisualizerModalClose}
+          initialState={visualizerModalInitialState}
+          saveLabel={t`Save`}
+        />
+      )}
     </ErrorBoundary>
   );
 }
