--- frontend/src/metabase/dashboard/components/DashCard/DashCardActionsPanel/VisualizerButton/index.ts
@@ -1,1 +1,0 @@
-export * from "./VisualizerButton";
